sql: hoist name resolution out of plan::query module
Eventually name resolution will be a pass that happens before planning.
Move it to the `names` module to start to make that clear.

This commit is pure code movement.
